+++ packages/engine/Source/Scene/ImageryLayer.js
@@ -1133,18 +1133,26 @@ ImageryLayer.prototype._requestImagery = function (imagery) {
     TileProviderError.reportSuccess(that._requestImageError);
   }
 
-  function failure(e) {
+  function failureWithoutRetry(e) {
     if (imagery.request.state === RequestState.CANCELLED) {
       // Cancelled due to low priority - try again later.
       imagery.state = ImageryState.UNLOADED;
       imagery.request = undefined;
-      return;
+      return true;
     }
 
     // Initially assume failure. An error handler may retry, in which case the state will
-    // change to TRANSITIONING.
+    // change to TRANSITIONING when doRequest is called.
     imagery.state = ImageryState.FAILED;
     imagery.request = undefined;
+
+    return false;
+  }
+
+  function failure(e) {
+    if (failureWithoutRetry(e)) {
+      return;
+    }
 
     const message = `Failed to obtain image tile X: ${imagery.x} Y: ${imagery.y} Level: ${imagery.level}.`;
     that._requestImageError = TileProviderError.reportError(
@@ -1158,11 +1166,11 @@ ImageryLayer.prototype._requestImagery = function (imagery) {
       e
     );
     if (that._requestImageError.retry) {
-      doRequest();
+      doRequest(Promise.resolve(that._requestImageError.retry));
     }
   }
 
-  function doRequest() {
+  function doRequest(waitPromise) {
     const request = new Request({
       throttle: false,
       throttleByServer: true,
@@ -1170,6 +1178,20 @@ ImageryLayer.prototype._requestImagery = function (imagery) {
     });
     imagery.request = request;
     imagery.state = ImageryState.TRANSITIONING;
+
+    if (waitPromise) {
+      // We need to wait for this promise to resolve before retrying.
+      // This promise might be, for example, obtaining a new token that will allow access to the imagery.
+      // We wrap doRequest in a function so that the resolved value of the waitPromise is ignored and
+      // not passed to doRequest.
+      waitPromise
+        .then(function () {
+          doRequest();
+        })
+        .catch(failureWithoutRetry);
+      return;
+    }
+
     const imagePromise = imageryProvider.requestImage(
       imagery.x,
       imagery.y,